+++ CHANGELOG.md
@@ -11,10 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support WordPress "Quote" block in content fields.
+- "Campo-Key" term meta field.
 
 ### Changed
 
-- Make "Duration of studies in semester" a single line input. 
+- Make "Duration of studies in semester" a single line input.
 
 ## [1.2.7] - 2023-11-23